fix: changed package name
